refactor: rename renderBook to createBookCard
- Create function createBookCard in dom.ts
- Import createBookCard in main.ts
- Replace use of renderBook with createBookCard
- Delete implementation of renderBook
